+++ src/utils/useNavigate.js
@@ -1,6 +1,6 @@
+import { store } from '../store';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
-import { store } from '../store';
 
 export const useNavigate = (routes) => {
   const navigate = (path) => {
